+++ packages/rax-view/src/index.tsx
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 import { createElement, useRef, forwardRef, ForwardRefExoticComponent, MutableRefObject } from 'rax';
 import cx from 'classnames/dedupe';
 import { isWeex } from 'universal-env';
@@ -7,6 +8,23 @@ import './index.css';
 
 interface ViewRef extends MutableRefObject<HTMLDivElement> {
   triggeredAppear: boolean;
+=======
+import { isMiniApp, isWeb, isWeChatMiniProgram } from 'universal-env';
+import ViewCommon from './common';
+import ViewAliMiniApp from './miniapp/ali';
+import ViewWechatMiniProgram from './miniapp/wechat';
+
+let View = null;
+
+if (isWeb) {
+  View = ViewCommon;
+} else if (isMiniApp) {
+  View = ViewAliMiniApp;
+} else if (isWeChatMiniProgram) {
+  View = ViewWechatMiniProgram;
+} else {
+  View = ViewCommon;
+>>>>>>> master
 }
 
 const View: ForwardRefExoticComponent<ViewProps> = forwardRef(