+++ packages/rax-countdown/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.1.7
+
+- Keep miniapp native code in older dir
+
 ## v1.1.6
 
 - adjust miniapp native code