+++ packages/rax-embed/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.2.2
+
+- Remove `sideEffects: false` in package.json
+
 ## 1.2.1
 
 - Add `sideEffects: false` in package.json