+++ packages/rax-video/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.5.2
+
+- Keep miniapp/wechat native code in older dir
+- Remove `sideEffects: false` in package.json
+
 ## 1.5.1
 
 - Add `sideEffects: false` in package.json